[#9] Refactor AdjustOne into a component.
