+++ src/css/adjustOne.css
@@ -1,0 +1,7 @@
+.adjustOne > span {
+  margin-right: 0.5em;
+}
+
+.adjustOneToLabel {
+  margin-left: 0.5em;
+}
